Extract account level tests
