+++ pkg/resources/secondary_database_acceptance_test.go
@@ -1,3 +1,5 @@
+//go:build account_level_tests
+
 package resources_test
 
 import (
@@ -20,7 +22,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// TODO [SNOW-1991414]: discuss and address all the nondeterministic tests in this file
 func TestAcc_CreateSecondaryDatabase_Basic(t *testing.T) {
 	_ = testenvs.GetOrSkipTest(t, testenvs.EnableAcceptance)
 	acc.TestAccPreCheck(t)
